refactor sponsorship.py to use new module, civicrm + small fixes
- remove unused imports
- remove stub functions
