+++ openlibrary/core/sponsorships.py
@@ -1,19 +1,18 @@
-import json
 import urllib
 import requests
-import web
 import logging
 from infogami import config
 from infogami.utils.view import public
-from openlibrary import accounts
 from openlibrary.core import models
 from openlibrary.core.lending import (
     get_work_availability, config_ia_domain)
 from openlibrary.core.vendors import (
-    get_betterworldbooks_metadata, create_edition_from_amazon_metadata,
+    get_betterworldbooks_metadata,
     get_amazon_metadata)
 from openlibrary.accounts import get_internet_archive_id
-from openlibrary.core.lending import config_ia_civicrm_api
+from openlibrary.core.civicrm import (
+    get_contact_id_by_username,
+    get_sponsorships_by_contact_id)
 from openlibrary.utils.isbn import to_isbn_13
 try:
     from booklending_utils.sponsorship import eligibility_check
@@ -23,10 +22,6 @@ except ImportError:
         return False
 
 logger = logging.getLogger("openlibrary.sponsorship")
-
-CIVI_ISBN = 'custom_52'
-CIVI_USERNAME = 'custom_51'
-CIVI_CONTEXT = 'custom_53'
 
 
 def get_sponsored_editions(user):
@@ -38,58 +33,6 @@ def get_sponsored_editions(user):
     contact_id = get_contact_id_by_username(archive_id)
     return contact_id and get_sponsorships_by_contact_id(contact_id)
 
-
-def get_contact_id_by_username(username):
-    """TODO: Use CiviCRM Explorer to replace with call to get contact_id by username"""
-    data = {
-        'entity': 'Contact',
-        'action': 'get',
-        'api_key': config_ia_civicrm_api.get('api_key', ''),
-        'key': config_ia_civicrm_api.get('site_key', ''),
-        'json': {
-            "sequential": 1,
-            CIVI_USERNAME: username
-        }
-    }
-    data['json'] = json.dumps(data['json'])  # flatten the json field as a string
-    r = requests.get(
-        config_ia_civicrm_api.get('url', ''),
-        params=urllib.urlencode(data),
-        headers={
-            'Authorization': 'Basic %s' % config_ia_civicrm_api.get('auth', '')
-        })
-    contacts = r.json().get('values', None)
-    return contacts and contacts[0].get('contact_id')
-
-
-def get_sponsorships_by_contact_id(contact_id, isbn=None):
-    data = {
-        'entity': 'Contribution',
-        'action': 'get',
-        'api_key': config_ia_civicrm_api.get('api_key', ''),
-        'key': config_ia_civicrm_api.get('site_key', ''),
-        'json': {
-            "sequential": 1,
-            "financial_type_id": "Book Sponsorship",
-            "contact_id": contact_id
-        }
-    }
-    if isbn:
-        data['json'][CIVI_ISBN] = isbn
-    data['json'] = json.dumps(data['json'])  # flatten the json field as a string
-    r = requests.get(
-        config_ia_civicrm_api.get('url', ''),
-        params=urllib.urlencode(data),
-        headers={
-            'Authorization': 'Basic %s' % config_ia_civicrm_api.get('auth', '')
-        })
-    txs = r.json().get('values')
-    return [{
-        'isbn': t.pop(CIVI_ISBN),
-        'context': t.pop(CIVI_CONTEXT),
-        'receive_date': t.pop('receive_date'),
-        'total_amount': t.pop('total_amount')
-    } for t in txs]
 
 def do_we_want_it(isbn, work_id):
     """Returns True if we don't have this edition (or other editions of
@@ -186,11 +129,3 @@ def qualifies_for_sponsorship(edition):
         })
     })
     return resp
-
-
-def get_all_sponsors():
-    """TODO: Query civi for a list of all sponsorships, for all users. These
-    results will have to be summed by user for the leader-board and
-    then cached
-    """
-    pass
